+++ src/__tests__/integration/google/request.test.ts
@@ -3,7 +3,7 @@ import fsp from 'fs/promises';
 
 import { createAccount } from '../../../account';
 import { fetchCalendarObjects, fetchCalendars } from '../../../calendar';
-import { DAVNamespace } from '../../../consts';
+import { DAVNamespaceShort } from '../../../consts';
 import { createObject, davRequest, deleteObject, propfind, updateObject } from '../../../request';
 import { DAVAccount, DAVCalendar } from '../../../types/models';
 import { getOauthHeaders } from '../../../util/authHelpers';
@@ -124,7 +124,9 @@ test('davRequest should be able to get raw xml response', async () => {
 test('propfind should be able to find props', async () => {
   const [result] = await propfind({
     url: 'https://apidata.googleusercontent.com/caldav/v2/',
-    props: [{ name: 'current-user-principal', namespace: DAVNamespace.DAV }],
+    props: {
+      [`${DAVNamespaceShort.DAV}:current-user-principal`]: {},
+    },
     depth: '0',
     headers: authHeaders,
   });